feat: got store_oracle working
