+++ scripts/prove/bin/multi.rs
@@ -36,10 +36,10 @@ async fn main() -> Result<()> {
         .get_host_cli_args(l2_start_block, l2_end_block, ProgramType::Multi, cache_mode)
         .await?;
 
-    let mem_kv_store = start_server_and_native_client(&host_cli).await?;
+    let oracle = start_server_and_native_client(&host_cli).await?;
 
     // Get the stdin for the block.
-    let sp1_stdin = get_proof_stdin(&host_cli, mem_kv_store)?;
+    let sp1_stdin = get_proof_stdin(oracle)?;
 
     let prover = ProverClient::from_env();
