+++ scripts/utils/bin/gen_sp1_test_artifacts.rs
@@ -56,8 +56,8 @@ async fn main() -> Result<()> {
 
     let mut successful_ranges = Vec::new();
     for (range, host_cli) in split_ranges.iter().zip(host_clis.iter()) {
-        let mem_kv_store = start_server_and_native_client(host_cli).await.unwrap();
-        let sp1_stdin = get_proof_stdin(host_cli, mem_kv_store).unwrap();
+        let oracle = start_server_and_native_client(host_cli).await.unwrap();
+        let sp1_stdin = get_proof_stdin(oracle).unwrap();
         successful_ranges.push((sp1_stdin, range.clone()));
     }
