+++ utils/client/src/lib.rs
@@ -7,10 +7,12 @@ pub mod boot;
 pub use boot::{BootInfoWithBytesConfig, AGGREGATION_OUTPUTS_SIZE};
 
 mod oracle;
-pub use oracle::InMemoryOracle;
+pub use oracle::{InMemoryOracle, StoreOracle};
 
 pub mod precompiles;
 
 pub mod types;
 
 extern crate alloc;
+
+pub mod client;
