+++ utils/host/src/lib.rs
@@ -5,27 +5,27 @@ pub mod rollup_config;
 pub mod stats;
 
 use alloy_consensus::Header;
-use alloy_primitives::{map::HashMap, B256};
+use alloy_primitives::B256;
 use alloy_sol_types::sol;
 use anyhow::anyhow;
 use anyhow::Result;
+use kona_host::single::SingleChainHostCli;
 use kona_host::HostOrchestrator;
 use kona_host::PreimageServer;
-use kona_host::{single::SingleChainHostCli, DiskKeyValueStore, MemoryKeyValueStore};
 use kona_preimage::BidirectionalChannel;
 use kona_preimage::HintReader;
 use kona_preimage::HintWriter;
 use kona_preimage::OracleReader;
 use kona_preimage::OracleServer;
 use log::info;
-use maili_genesis::RollupConfig;
+use op_succinct_client_utils::client::run_opsuccinct_client;
+use op_succinct_client_utils::precompiles::zkvm_handle_register;
 use op_succinct_client_utils::InMemoryOracle;
-use op_succinct_client_utils::{
-    boot::BootInfoStruct, types::AggregationInputs, BootInfoWithBytesConfig, BytesHasherBuilder,
-};
+use op_succinct_client_utils::StoreOracle;
+use op_succinct_client_utils::{boot::BootInfoStruct, types::AggregationInputs};
 use rkyv::to_bytes;
 use sp1_sdk::{HashableKey, SP1Proof, SP1Stdin};
-use std::{fs::File, io::Read};
+use std::sync::Arc;
 use tokio::task;
 
 sol! {
@@ -58,43 +58,11 @@ sol! {
 }
 
 /// Get the stdin to generate a proof for the given L2 claim.
-pub fn get_proof_stdin(
-    host_cli: &SingleChainHostCli,
-    mem_kv_store: MemoryKeyValueStore,
-) -> Result<SP1Stdin> {
+pub fn get_proof_stdin(oracle: InMemoryOracle) -> Result<SP1Stdin> {
     let mut stdin = SP1Stdin::new();
 
-    // Read the rollup config.
-    let mut rollup_config_file = File::open(host_cli.rollup_config_path.as_ref().unwrap())?;
-    let mut rollup_config_bytes = Vec::new();
-    rollup_config_file.read_to_end(&mut rollup_config_bytes)?;
-
-    let ser_config = std::fs::read_to_string(host_cli.rollup_config_path.as_ref().unwrap())?;
-    let rollup_config: RollupConfig = serde_json::from_str(&ser_config)?;
-
-    let boot_info = BootInfoWithBytesConfig {
-        l1_head: host_cli.l1_head,
-        l2_output_root: host_cli.agreed_l2_output_root,
-        l2_claim: host_cli.claimed_l2_output_root,
-        l2_claim_block: host_cli.claimed_l2_block_number,
-        chain_id: rollup_config.l2_chain_id,
-        rollup_config_bytes,
-    };
-    stdin.write(&boot_info);
-
-    // Convert the memory KV store to a HashMap<[u8;32], Vec<u8>>.
-    let mut kv_store_map: HashMap<[u8; 32], Vec<u8>, BytesHasherBuilder> =
-        HashMap::with_hasher(BytesHasherBuilder);
-    for (k, v) in mem_kv_store.store {
-        kv_store_map.insert(k.0, v);
-    }
-
-    let in_memory_oracle = InMemoryOracle {
-        cache: kv_store_map,
-    };
-
     // Serialize the underlying KV store.
-    let buffer = to_bytes::<rkyv::rancor::Error>(&in_memory_oracle)?;
+    let buffer = to_bytes::<rkyv::rancor::Error>(&oracle)?;
 
     let kv_store_bytes = buffer.into_vec();
     stdin.write_slice(&kv_store_bytes);
@@ -134,7 +102,7 @@ pub fn get_agg_proof_stdin(
 /// TODO: Can we run many program tasks in parallel?
 pub async fn start_server_and_native_client(
     cfg: &SingleChainHostCli,
-) -> Result<MemoryKeyValueStore, anyhow::Error> {
+) -> Result<InMemoryOracle, anyhow::Error> {
     let hint_chan = BidirectionalChannel::new().map_err(|e| anyhow!(e))?;
     let preimage_chan = BidirectionalChannel::new().map_err(|e| anyhow!(e))?;
 
@@ -152,27 +120,39 @@ pub async fn start_server_and_native_client(
         .start(),
     );
 
-    let program_task = tokio::spawn(SingleChainHostCli::run_client_native(
-        HintWriter::new(hint_chan.client),
+    ////////////////////////////////////////////////////////////////
+    //                          PROLOGUE                          //
+    ////////////////////////////////////////////////////////////////
+
+    // TODO: Switch back to caching oracle once we create a struct that wraps it that can save data.
+    // const LRU_SIZE: usize = 1024_1024_1024;
+    // let oracle = Arc::new(CachingOracle::new(
+    //     LRU_SIZE,
+    //     OracleReader::new(preimage_chan.client),
+    //     HintWriter::new(hint_chan.client),
+    // ));
+    let oracle = Arc::new(StoreOracle::new(
         OracleReader::new(preimage_chan.client),
+        HintWriter::new(hint_chan.client),
     ));
 
+    let program_task = task::spawn(run_opsuccinct_client(
+        oracle.clone(),
+        Some(zkvm_handle_register),
+    ));
+
+    // TODO: Clean this up.
     info!("Starting preimage server and client program.");
-    let (_, client_result) =
-        tokio::try_join!(server_task, program_task,).map_err(|e| anyhow!(e))?;
-    if let Err(e) = client_result {
-        return Err(e);
-    }
-    info!(target: "kona_host", "Preimage server and client program have joined.");
+    let _ = tokio::select! {
+        r = server_task => {
+            r?;
+            return Err(anyhow!("Server task completed before program task"));
+        },
+        r = program_task => r??,
+    };
 
-    // Drop the KV store after the server is complete, to avoid conflicting locks.
-    drop(kv_store);
-
-    // Loop over all of the keys in the KV store and convert to MemoryKeyValueStore.
-    let disk_kv_store = DiskKeyValueStore::new(cfg.data_dir.clone().unwrap());
-
-    let mem_kv_store: MemoryKeyValueStore = disk_kv_store.try_into()?;
+    let in_memory_oracle = InMemoryOracle::populate_from_store(&oracle).unwrap();
 
     // Convert to MemoryKeyValueStore and return it
-    Ok(mem_kv_store)
+    Ok(in_memory_oracle)
 }
